nonfunctional_updates_to_satisfy_google_cpplint
