+++ src/freeserf.cc
@@ -46,7 +46,7 @@ main(int argc, char *argv[]) {
   new(Log);
   std::ofstream* filestr = new std::ofstream("console_out.txt");
   Log::set_file(filestr);
-  
+
   std::string data_dir;
   std::string save_file;
 
@@ -109,12 +109,12 @@ main(int argc, char *argv[]) {
 
   Graphics &gfx = Graphics::get_instance();
 
-  /* TODO move to right place */
-    Audio &audio = Audio::get_instance();
-    Audio::PPlayer player = audio.get_music_player();
-    if (player) {
-      Audio::PTrack t = player->play_track(Audio::TypeMidiTrack0);
-    }
+  ///* TODO move to right place */
+  //  Audio &audio = Audio::get_instance();
+  //  Audio::PPlayer player = audio.get_music_player();
+  //  if (player) {
+  //    Audio::PTrack t = player->play_track(Audio::TypeMidiTrack0);
+  //  }
 
   GameManager &game_manager = GameManager::get_instance();
 
@@ -144,7 +144,6 @@ main(int argc, char *argv[]) {
   }
   else {
 	  loaded_game_start_ai = true;
-    //interface.initialize_AI();
   }
 
   /* Init game loop */
